requirements.txt
Flask==3.0.1
Flask-Cors==4.0.0
APScheduler==3.10.4
python-dotenv==1.0.1
requests==2.31.0
Pillow==10.0.0
matplotlib==3.7.2
pandas==2.0.3
PyMuPDF==1.24.10
pdf2image==1.16.3
openpyxl==3.1.2
Werkzeug==3.0.1
numpy==1.25.1
schedule==1.2.0
pytz==2023.3
pymsteams==0.2.5